Split div struct BDD steps (#60)
Move the div-specific `msgspec.Struct` round-trip assertions and scenario
binding into `steps_div_structs.rs`. Keep `steps_structs.rs` focused on the
generic Episode struct steps so both modules remain under the repository file
size limit.
